refactor(enum): 패키지 명 `enum` → `enumerator`로 변경
