+++ src/main/java/lotto/domain/LottoResult.java
@@ -1,7 +1,7 @@
 package lotto.domain;
 
 import java.util.List;
-import lotto.util.enums.LottoRank;
+import lotto.util.enumerator.LottoRank;
 
 public class LottoResult {
     private final List<LottoRank> rankResults;
